Change DialogueManager and ChoiceManager to implement singletons
If changing methods in both managers to static methods, non-static
fields such as dialogBox in DialogueManager can not be called.
Singletons without DontDestryOnLoad() is used as managers are not
preserved through scenes, this is to make possible for future
alterations of managers such as aesthetics in certain scenes.

Interactable: '&& !GameManager.instance.playerFrozen' is added to
line 22 as without it bugs occurs when the player is in the trigger zone
of a Interactable object and press 'Z' in with the inventory open.

Item: 'amount--' commented out as most items are not usable from
the inventory. When the player tries to use the item, it will show the
zoom in image and description instead. To implement usable items such as
rubber balls we can extend Item with a child class and override the
Use() method.