Change the mechanism of fishing
Previously, Fishing.cs is used which calls Update() in which gameEvent
is called. This is not very ideal. Delete Fishing.cs and call
startFishing and endFishing event directly from FishingArea.

Change FishingArea and PhoenixFish so that the success or failure
conversation are both triggered in FishingArea when it does the check,
while the sole functionality of PhoenixFish is to add the fish into the
inventory which FishingArea can check in its CheckIfHaveFish Coroutine.

Ui: add the missing field of Name(of the item) GameObject to ZoomInBox