Refactor the inventory system
Make Item abstract and extend it with ConsumableItem, Specialitem and
DescriptiveItem

SpecialItem: can be used directly from the inventory which behaves the
same as the player interacts with the Interactable Object with the
required item, e.g. passing the hair pin to Eri can be implemented by
both directly interacting with Eri or open the inventory and chooses the
hair pin.

DescriptiveItem: cannot be used anywhere and can only be used to see the
zoom in description.

Refactored Follower and changed Follower_simple to SimpleFollower.
Changed phoenixFish(which used to extend Follower_simple) accordingly.

GameEvent is used in SpecialItem to perform the effect of the item.
Extends GameEvent with SpecialItemGameEvent that closes the inventory UI
after calling the usual TriggerEvent. If not, the inventory UI and the
dialogue box may exist at the same time as some effects of the item
triggers dialogues.