Fix bugs related to merging
Change tag so that the tag of the demi-human and human kizuna to
"Player", the spirit to "Spirit" and the partner remain "Partner". Edit
the CameraManager accordingly.

CameraManager: make the label = "main" by default so that it is not
needed to change it in the inspector for every scene

CharacterMovement: change AlterLocationType(bool isContinuousType) to
non-static, otherwise will not be able to call it in the inspector.

Castle map bug: fixed by initiating dictionaries and calling
AlterLocationType(bool isContinuousType) in the timeline in
Hall_Before_Chase cutscene instead of using GameEvent objects.